HTRACE-314. htraced: make datastore loading safer (cmccabe)

+++ htrace-htraced/go/src/org/apache/htrace/htraced/mini_htraced.go
@@ -115,8 +115,10 @@ func (bld *MiniHTracedBuilder) Build() (*MiniHTraced, error) {
 				store.Close()
 			}
 			for idx := range bld.DataDirs {
-				if bld.DataDirs[idx] != "" {
-					os.RemoveAll(bld.DataDirs[idx])
+				if !bld.KeepDataDirsOnClose {
+					if bld.DataDirs[idx] != "" {
+						os.RemoveAll(bld.DataDirs[idx])
+					}
 				}
 			}
 			if rsv != nil {
